+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraTargetedSweepIntegrationTest.java
@@ -15,70 +15,21 @@
  */
 package com.palantir.atlasdb.keyvalue.cassandra;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-import java.nio.charset.StandardCharsets;
-import java.util.Optional;
-
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.rules.RuleChain;
 
-import com.google.common.collect.ImmutableMap;
-import com.palantir.atlasdb.containers.CassandraContainer;
 import com.palantir.atlasdb.containers.CassandraResource;
-import com.palantir.atlasdb.keyvalue.api.Cell;
-import com.palantir.atlasdb.keyvalue.api.KeyValueService;
-import com.palantir.atlasdb.keyvalue.api.SweepResults;
-import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.sweep.AbstractSweepTest;
-import com.palantir.atlasdb.sweep.queue.ShardAndStrategy;
-import com.palantir.atlasdb.sweep.queue.SpecialTimestampsSupplier;
-import com.palantir.atlasdb.sweep.queue.TargetedSweepFollower;
-import com.palantir.atlasdb.sweep.queue.TargetedSweeper;
-import com.palantir.lock.v2.TimelockService;
+import com.palantir.atlasdb.sweep.AbstractTargetedSweepTest;
 
-public class CassandraTargetedSweepIntegrationTest extends AbstractSweepTest {
-    private SpecialTimestampsSupplier timestampsSupplier = mock(SpecialTimestampsSupplier.class);
-    private TargetedSweeper sweepQueue;
-
+public class CassandraTargetedSweepIntegrationTest extends AbstractTargetedSweepTest {
     @ClassRule
-    public static final CassandraResource CASSANDRA = new CassandraResource(
-            CassandraTargetedSweepIntegrationTest.class);
+    public static final CassandraResource CASSANDRA = new CassandraResource();
 
     @Rule
-    public final RuleChain ruleChain = SchemaMutationLockReleasingRule.createChainedReleaseAndRetry(
-            getKeyValueService(), CASSANDRA.getConfig());
+    public final RuleChain ruleChain = SchemaMutationLockReleasingRule.createChainedReleaseAndRetry(CASSANDRA);
 
-    @Before
-    public void setup() {
-        super.setup();
-
-        sweepQueue = TargetedSweeper.createUninitializedForTest(() -> 1);
-        sweepQueue.initializeWithoutRunning(
-                timestampsSupplier, mock(TimelockService.class), kvs, mock(TargetedSweepFollower.class));
-    }
-
-    @Override
-    protected KeyValueService getKeyValueService() {
-        return CassandraKeyValueServiceImpl
-                .createForTesting(CASSANDRA.getConfig(), CassandraContainer.LEADER_CONFIG);
-    }
-
-    @Override
-    protected Optional<SweepResults> completeSweep(TableReference tableReference, long ts) {
-        when(timestampsSupplier.getUnreadableTimestamp()).thenReturn(ts);
-        when(timestampsSupplier.getImmutableTimestamp()).thenReturn(ts);
-        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(0));
-        sweepQueue.sweepNextBatch(ShardAndStrategy.thorough(0));
-        return Optional.empty();
-    }
-
-    @Override
-    protected void put(final TableReference tableRef, Cell cell, final String val, final long ts) {
-        super.put(tableRef, cell, val, ts);
-        sweepQueue.enqueue(ImmutableMap.of(tableRef, ImmutableMap.of(cell, val.getBytes(StandardCharsets.UTF_8))), ts);
+    public CassandraTargetedSweepIntegrationTest() {
+        super(CASSANDRA, CASSANDRA);
     }
 }