+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/Schema.java
@@ -15,6 +15,12 @@
  */
 package com.palantir.atlasdb.table.description;
 
+import static java.util.stream.Collectors.toSet;
+
+import static com.google.common.collect.ImmutableSet.toImmutableSet;
+import static com.google.common.collect.Maps.immutableEntry;
+import static com.palantir.logsafe.Preconditions.checkState;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -38,6 +44,7 @@ import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Collections2;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.ListMultimap;
@@ -45,12 +52,14 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Ordering;
+import com.google.common.collect.SetMultimap;
 import com.google.common.collect.Sets;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.cleaner.api.OnCleanupTask;
 import com.palantir.atlasdb.keyvalue.api.Namespace;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.impl.AbstractKeyValueService;
+import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.LogSafety;
 import com.palantir.atlasdb.schema.ImmutableSchemaDependentTableMetadata;
 import com.palantir.atlasdb.schema.ImmutableSchemaMetadata;
 import com.palantir.atlasdb.schema.SchemaDependentTableMetadata;
@@ -67,6 +76,9 @@ import com.palantir.atlasdb.table.description.render.TableFactoryRenderer;
 import com.palantir.atlasdb.table.description.render.TableRenderer;
 import com.palantir.atlasdb.table.description.render.TableRendererV2;
 import com.palantir.atlasdb.transaction.api.ConflictHandler;
+import com.palantir.common.streams.KeyedStream;
+import com.palantir.logsafe.SafeArg;
+import com.palantir.logsafe.UnsafeArg;
 
 /**
  * Defines a schema.
@@ -99,6 +111,8 @@ public class Schema {
     // Used to determine cleanup metadata.
     private final Set<String> tablesWithCustomCleanupTasks = Sets.newHashSet();
     private final Map<String, StreamStoreCleanupMetadata> streamStoreCleanupMetadata = Maps.newHashMap();
+
+    private final Set<String> deprecatedTableNames = Sets.newHashSet();
 
     /** Creates a new schema, using Guava Optionals. */
     public Schema(Namespace namespace) {
@@ -161,14 +175,14 @@ public class Schema {
 
     public Set<TableReference> getAllIndexes() {
         return indexDefinitions.keySet().stream()
-                .map((table) -> TableReference.create(namespace, table))
-                .collect(Collectors.toSet());
+                .map(table -> TableReference.create(namespace, table))
+                .collect(toSet());
     }
 
     public Set<TableReference> getAllTables() {
         return tableDefinitions.keySet().stream()
-                .map((table) -> TableReference.create(namespace, table))
-                .collect(Collectors.toSet());
+                .map(table -> TableReference.create(namespace, table))
+                .collect(toSet());
     }
 
     public void addIndexDefinition(String idxName, IndexDefinition definition) {
@@ -199,6 +213,24 @@ public class Schema {
                             + "If running only against a different KVS, set the ignoreTableNameLength flag.",
                     idxName, StringUtils.join(kvsExceeded, ", "));
         }
+    }
+
+    /**
+     * Registers old tables referred to by {@code deprecatedTableNames} to be dropped from backing store on startup.
+     *
+     * If running in a multi-node configuration, you must ensure that tables added in one version are not being
+     * read/written to in the other running version to avoid consistency problems.
+     *
+     * @param deprecatedTableNames old tables that should be dropped
+     */
+    public void addDeprecatedTables(String... deprecatedTableNames) {
+        this.deprecatedTableNames.addAll(ImmutableList.copyOf(deprecatedTableNames));
+    }
+
+    public Set<TableReference> getDeprecatedTables() {
+        return deprecatedTableNames.stream()
+                .map(tableName -> TableReference.create(namespace, tableName))
+                .collect(toImmutableSet());
     }
 
     /**
@@ -303,6 +335,25 @@ public class Schema {
                         "Nonadditive indexes require write-write conflicts on their tables");
             }
         }
+
+        validateDeprecatedTables();
+    }
+
+    private void validateDeprecatedTables() {
+        Map<TableReference, TableMetadata> allTablesAndIndexMetadata = getAllTablesAndIndexMetadata();
+        Set<TableReference> invalidTables =
+                Sets.intersection(allTablesAndIndexMetadata.keySet(), getDeprecatedTables());
+
+        SetMultimap<LogSafety, TableReference> referencesByLogSafety = KeyedStream.stream(
+                allTablesAndIndexMetadata)
+                .filterKeys(invalidTables::contains)
+                .mapEntries((reference, metadata) -> immutableEntry(metadata.getNameLogSafety(), reference))
+                .collectToSetMultimap();
+
+        checkState(invalidTables.isEmpty(),
+                "A deprecated table cannot also be part of your schema. Check logs for any unsafe table names.",
+                SafeArg.of("invalidDeprecatedTables_safe", referencesByLogSafety.get(LogSafety.SAFE)),
+                UnsafeArg.of("invalidDeprecatedTables_unsafe", referencesByLogSafety.get(LogSafety.UNSAFE)));
     }
 
     public Map<TableReference, TableDefinition> getTableDefinitions() {