+++ timelock-agent/src/main/java/com/palantir/timelock/paxos/PaxosLeadershipCreator.java
@@ -21,6 +21,7 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
+import com.google.common.base.Suppliers;
 import com.palantir.atlasdb.config.ImmutableLeaderConfig;
 import com.palantir.atlasdb.config.ImmutableLeaderRuntimeConfig;
 import com.palantir.atlasdb.config.LeaderConfig;
@@ -38,7 +39,6 @@ import com.palantir.leader.proxy.AwaitingLeadershipProxy;
 import com.palantir.timelock.config.PaxosRuntimeConfiguration;
 import com.palantir.timelock.config.TimeLockInstallConfiguration;
 import com.palantir.timelock.config.TimeLockRuntimeConfiguration;
-import com.palantir.util.JavaSuppliers;
 
 public class PaxosLeadershipCreator {
     private final MetricsManager metricsManager;
@@ -57,7 +57,7 @@ public class PaxosLeadershipCreator {
             Consumer<Object> registrar) {
         this.metricsManager = metricsManager;
         this.install = install;
-        this.runtime = JavaSuppliers.compose(TimeLockRuntimeConfiguration::paxos, runtime);
+        this.runtime = Suppliers.compose(TimeLockRuntimeConfiguration::paxos, runtime::get);
         this.registrar = registrar;
     }
 
@@ -71,7 +71,7 @@ public class PaxosLeadershipCreator {
         Leaders.LocalPaxosServices localPaxosServices = Leaders.createInstrumentedLocalServices(
                 metricsManager,
                 leaderConfig,
-                JavaSuppliers.compose(getLeaderRuntimeConfig, runtime),
+                Suppliers.compose(getLeaderRuntimeConfig::apply, runtime::get),
                 ImmutableRemotePaxosServerSpec.builder()
                         .remoteLeaderUris(remoteServers)
                         .remoteAcceptorUris(paxosSubresourceUris)